+++ mycodo/outputs/pcf8574.py
@@ -217,18 +217,22 @@ class OutputModule(AbstractOutput):
         self.setup_on_off_output(OUTPUT_INFORMATION)
 
         if self.output.i2c_location:
-            self.sensor = PCF8574(smbus2, self.output.i2c_bus, int(str(self.output.i2c_location), 16))
+            # self.sensor = PCF8574(smbus2, self.output.i2c_bus, int(str(self.output.i2c_location), 16))
             self.output_setup = True
 
+        dict_states = {}
         for channel in channels_dict:
             if self.options_channels['state_startup'][channel] == 1:
-                self.output_switch('on', output_channel=channel)
+                dict_states[channel] = bool(self.options_channels['on_state'][channel])
             elif self.options_channels['state_startup'][channel] == 0:
-                self.output_switch('off', output_channel=channel)
+                dict_states[channel] = bool(not self.options_channels['on_state'][channel])
             else:
                 continue
-            if self.options_channels['trigger_functions_startup'][channel]:
-                self.check_triggers(self.unique_id, output_channel=channel)
+
+        # self.sensor.port = list_states
+        self.output_states = dict_states
+        if self.options_channels['trigger_functions_startup'][channel]:
+            self.check_triggers(self.unique_id, output_channel=channel)
 
     def output_switch(self,
                       state,
@@ -241,17 +245,16 @@ class OutputModule(AbstractOutput):
             return
 
         try:
-            list_states = []
+            dict_states = {}
             for channel in channels_dict:
                 if output_channel == channel:
                     if state == 'on':
-                        list_states.append(self.options_channels['on_state'][channel])
+                        dict_states[channel] = bool(self.options_channels['on_state'][channel])
                     elif state == 'off':
-                        list_states.append(not self.options_channels['on_state'][channel])
-                else:
-                    list_states.append(self.output_states[channel])
-            self.sensor.port = list_states
-            self.output_states[output_channel] = list_states[output_channel]
+                        dict_states[channel] = bool(not self.options_channels['on_state'][channel])
+
+            # self.sensor.port = list_states
+            self.output_states[output_channel] = dict_states[output_channel]
             msg = "success"
         except Exception as e:
             msg = "CH{} state change error: {}".format(output_channel, e)
@@ -270,42 +273,14 @@ class OutputModule(AbstractOutput):
 
     def stop_output(self):
         """ Called when Output is stopped """
+        list_states = [0] * len(channels_dict)
         for channel in channels_dict:
             if self.options_channels['state_shutdown'][channel] == 1:
-                self.output_switch('on', output_channel=channel)
+                list_states[channel] = bool(self.options_channels['on_state'][channel])
             elif self.options_channels['state_shutdown'][channel] == 0:
-                self.output_switch('off', output_channel=channel)
+                list_states[channel] = bool(not self.options_channels['on_state'][channel])
+        self.sensor.port = list_states
         self.running = False
-
-
-class IOPort(list):
-    """ Represents the PCF8574 IO port as a list of boolean values """
-    def __init__(self, pcf8574, *args, **kwargs):
-        super(IOPort, self).__init__(*args, **kwargs)
-        self.pcf8574 = pcf8574
-
-    def __setitem__(self, key, value):
-        """ Set an individual output pin """
-        self.pcf8574.set_output(key, value)
-
-    def __repr__(self):
-        """ Represent port as a list of booleans """
-        state = self.pcf8574.bus.read_byte(self.pcf8574.address)
-        ret = []
-        for i in range(8):
-            ret.append(bool(state & 1 << i))
-        return repr(ret)
-
-    def __len__(self):
-        return 8
-
-    def __iter__(self):
-        for i in range(8):
-            yield self[i]
-
-    def __reversed__(self):
-        for i in range(8):
-            yield self[7-i]
 
 
 class PCF8574(object):
@@ -318,12 +293,6 @@ class PCF8574(object):
     def __repr__(self):
         return "PCF8574(i2c_bus_no=%r, address=0x%02x)" % (self.bus_no, self.address)
 
-    @property
-    def port(self):
-        """ Represent IO port as a list of boolean values """
-        return IOPort(self)
-
-    @port.setter
     def port(self, value):
         """ Set the whole port using a list """
         assert isinstance(value, list)
@@ -332,12 +301,4 @@ class PCF8574(object):
         for i, val in enumerate(value):
             if val:
                 new_state |= 1 << i
-        self.bus.write_byte(self.address, new_state)
-
-    def set_output(self, output_number, value):
-        """ Set a specific output high (True) or low (False) """
-        assert output_number in range(8), "Output number must be an integer between 0 and 7"
-        current_state = self.bus.read_byte(self.address)
-        bit = 1 << output_number
-        new_state = current_state | bit if value else current_state & (~bit & 0xff)
         self.bus.write_byte(self.address, new_state)